project funding total and total budget percentage calculation added

+++ applications/Unity.GrantManager/src/Unity.GrantManager.Application.Contracts/GrantApplications/CreateUpdateProjectInfoDto.cs
@@ -10,8 +10,8 @@ namespace Unity.GrantManager.GrantApplications
         public decimal? TotalProjectBudget { get; set; }
         public DateTime? ProjectStartDate { get; set; }
         public DateTime? ProjectEndDate { get; set; }
-        public float? PercentageTotalProjectBudget { get; set; }
-        public double? ProjectFundingTotal { get; set; }
+        public double? PercentageTotalProjectBudget { get; set; }
+        public decimal? ProjectFundingTotal { get; set; }
         public string? Community { get; set; }
         public int? CommunityPopulation { get; set; }
         public string? Acquisition { get; set; }
