+++ applications/Unity.GrantManager/src/Unity.GrantManager.Application.Contracts/GrantApplications/GrantApplicationDto.cs
@@ -40,8 +40,8 @@ public class GrantApplicationDto : AuditedEntityDto<Guid>
     public GrantApplicationState StatusCode { get; set; }
     public DateTime? ProjectStartDate { get; set; }
     public DateTime? ProjectEndDate { get; set; }
-    public float? PercentageTotalProjectBudget { get; set; }
-    public double? ProjectFundingTotal { get; set; }
+    public double? PercentageTotalProjectBudget { get; set; }
+    public decimal? ProjectFundingTotal { get; set; }
     public string? Community { get; set; }
     public int? CommunityPopulation { get; set; }
     public string? Acquisition { get; set; }
